ISSUE #2243 - move meta out of model helper
