+++ backend/models/meta.js
@@ -20,7 +20,9 @@ const mongoose = require("mongoose");
 const ModelFactory = require("./factory/modelFactory");
 const FileRef = require("./fileRef");
 const History = require("./history");
+const ModelSetting = require("./modelSetting");
 const Ref = require("./ref");
+const Scene = require("./scene");
 const Schema = mongoose.Schema;
 const db = require("../handler/db");
 const responseCodes = require("../response_codes.js");
@@ -201,6 +203,127 @@ const Meta = ModelFactory.createClass(
 		return `${arg.model}.scene`;
 	}
 );
+
+Meta.getMetadata = function(account, model, id) {
+	const projection = {
+		shared_id: 0,
+		paths: 0,
+		type: 0,
+		api: 0,
+		parents: 0
+	};
+
+	return Scene.findOne({account, model}, { _id: utils.stringToUUID(id) }, projection).then(obj => {
+		if(obj) {
+			return obj;
+		} else {
+			return Promise.reject(responseCodes.METADATA_NOT_FOUND);
+		}
+	});
+};
+
+Meta.getAllIdsWithMetadataField = function(account, model, branch, rev, fieldName, username) {
+	// Get the revision object to find all relevant IDs
+	let history;
+	let fullFieldName = "metadata";
+
+	if (fieldName && fieldName.length > 0) {
+		fullFieldName += "." + fieldName;
+	}
+
+	return History.getHistory({ account, model }, branch, rev).then(_history => {
+		history = _history;
+		if(!history) {
+			return Promise.reject(responseCodes.METADATA_NOT_FOUND);
+		}
+		// Check for submodel references
+		const filter = {
+			type: "ref",
+			_id: { $in: history.current }
+		};
+		return Ref.find({ account, model }, filter);
+	}).then(refs =>{
+
+		// for all refs get their tree
+		const getMeta = [];
+
+		refs.forEach(ref => {
+
+			let refBranch, refRev;
+
+			if (utils.uuidToString(ref._rid) === C.MASTER_BRANCH) {
+				refBranch = C.MASTER_BRANCH_NAME;
+			} else {
+				refRev = utils.uuidToString(ref._rid);
+			}
+
+			getMeta.push(
+				getAllIdsWithMetadataField(ref.owner, ref.project, refBranch, refRev, fieldName, username)
+					.then(obj => {
+						return Promise.resolve({
+							data: obj.data,
+							account: ref.owner,
+							model: ref.project
+						});
+					})
+					.catch(() => {
+					// Just because a sub model fails doesn't mean everything failed. Resolve the promise.
+						return Promise.resolve();
+					})
+			);
+		});
+
+		return Promise.all(getMeta);
+
+	}).then(_subMeta => {
+
+		const match = {
+			_id: {"$in": history.current}
+		};
+		match[fullFieldName] =  {"$exists" : true};
+
+		const projection = {
+			parents: 1
+		};
+		projection[fullFieldName] = 1;
+
+		return Scene.find({account, model}, match, projection).then(obj => {
+			if(obj) {
+				// rename fieldName to "value"
+				const parsedObj = {data: obj};
+				if(obj.length > 0 && fieldName && fieldName.length > 0) {
+					const objStr = JSON.stringify(obj);
+					parsedObj.data = JSON.parse(objStr.replace(new RegExp(fieldName, "g"), "value"));
+				}
+				if(_subMeta.length > 0) {
+					parsedObj.subModels = _subMeta;
+				}
+				return parsedObj;
+			} else {
+				return Promise.reject(responseCodes.METADATA_NOT_FOUND);
+			}
+		});
+
+	});
+};
+
+Meta.getAllMetadata = function(account, model, branch, rev) {
+	return getAllIdsWithMetadataField(account, model, branch, rev, "");
+};
+
+Meta.getAllIdsWith4DSequenceTag = function(account, model, branch, rev) {
+	// Get sequence tag then call the generic getAllIdsWithMetadataField
+	return ModelSetting.findOne({account : account}, {_id : model}).then(settings => {
+		if(!settings) {
+			return Promise.reject(responseCodes.MODEL_NOT_FOUND);
+		}
+		if(!settings.fourDSequenceTag) {
+			return Promise.reject(responseCodes.SEQ_TAG_NOT_FOUND);
+		}
+		return getAllIdsWithMetadataField(account, model,  branch, rev, settings.fourDSequenceTag);
+
+	});
+};
 
 Meta.getMetadataFields = function(account, model) {
